Implement exact storage to ACTUALLY prevent rounding errors

+++ migration.js
@@ -4,6 +4,6 @@ db.prepare('DROP TABLE IF EXISTS Envelope').run();
 db.prepare(`CREATE TABLE Envelope (
     id INTEGER PRIMARY KEY,
     title TEXT NOT NULL,
-    budget NUMERIC NOT NULL,
+    budget INTEGER NOT NULL,
     CHECK(budget >= 0)
     )`).run();
